refactor(dry): Batch 2 — delete the never-invoked consolidators + sql_builders
The six consolidator classes (Metrics/Memory/Task/Conversation/Trace/
Audit) were constructed in TSDBConsolidationService.__init__ but never
invoked — the live consolidation path calls the substrate directly
(tsdb_consolidate_tasks/conversations/traces/audit +
telemetry_consolidate_period do the SQL, period aggregation, and
TEMPORAL_NEXT edges in-substrate). De-wired the imports + six
constructions and deleted consolidators/.

sql_builders.py: every build_* query builder is superseded by the same
substrate calls; the one live symbol, parse_datetime_field, moved to the
(live) date_calculation_helpers.py with its import updated in service.py.

Tests: removed the consolidators' 2 test files + test_sql_builders.py
(tests-only coverage of deleted code). Targeted graph-services suite:
318 passed.
