+++ ciris_engine/logic/services/graph/tsdb_consolidation/__init__.py
@@ -6,7 +6,6 @@ for long-term memory retention.
 
 Components:
 - service.py: Main service class and lifecycle management
-- consolidators/: Individual consolidation strategies by data type
 - edge_manager.py: Proper edge creation and management
 - query_manager.py: Querying nodes and correlations for consolidation
 - period_manager.py: Time period calculations and management
