--- tests/ciris_engine/logic/services/graph/tsdb_consolidation/consolidators/__init__.py
@@ -1,1 +1,0 @@
-# Tests for consolidators
